+++ predicate/Cargo.toml
@@ -13,6 +13,7 @@ generated_types = { path = "../generated_types" }
 internal_types = { path = "../internal_types" }
 observability_deps = { path = "../observability_deps" }
 regex = "1"
+serde_json = "1.0.67"
 snafu = "0.6.9"
 sqlparser = "0.11.0"
